backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2
psycopg2-binary==2.9.9; python_version < "3.13"
psycopg[binary]==3.2.10; python_version >= "3.13"
alembic==1.13.2
openpyxl==3.1.2
cloudinary==1.41.0
python-multipart==0.0.9
